triplet selection and triplet embedding function loss correct: positive, negative and anchor

+++ TripleData.lua
@@ -34,31 +34,182 @@ function LoadNormalizedResolutionImage(filename)
     return preprocess(imagepath)
 end
 
+
+function dist(a, b)
+    local d = (a -b)*(a-b)
+    return d
+    --return torch.dist(a,b)
+end
+
+function SelectListTriplets(embedding_net, db, size, TensorType)
+    print ("select list triplets", size)
+
+    local data = db.data
+    print ("dbsize:", #data.anchor_name_list)
+    local list = {}
+    local nClasses = #data.anchor_name_list 
+
+    local candidate_anchor_list = {}
+    for i=1, size do
+        local anchor_name = data.anchor_name_list[i]
+        if anchor_name ~= nil then
+            pos_of_anchor = data.positive[anchor_name]
+            if pos_of_anchor ~= nil then
+                print ("insert", anchor_name)
+                table.insert(candidate_anchor_list, anchor_name)
+            end
+        end
+    end
+
+    print ("canidate list", #candidate_anchor_list)
+
+    for i=1, size do
+        local anchor_img
+        local anchor_vector, positive_vector, negative_vector
+
+        local ap_dist, an_dist
+        print ("generate list #" .. i .. "/#" .. #candidate_anchor_list)
+        local c1, anchor_name, hard_positive_name, semi_hard_negative_name
+        c1 = math.random(#candidate_anchor_list)
+
+        local nsz = torch.LongStorage(4)
+        while true do
+            anchor_name = candidate_anchor_list[c1]
+            local batch = torch.Tensor():type( TensorType )
+            --print ( 'anchor_name:', anchor_name )
+
+            anchor_img = LoadNormalizedResolutionImage(anchor_name)
+            nsz[1] = 1
+            nsz[2] = 3
+            nsz[3] = 299
+            nsz[4] = 299
+            batch:resize(nsz)
+            batch[1]:copy(anchor_img)
+
+            --print ("anchor_image:size", anchor_img:size())
+            a_output = embedding_net:forward( batch )
+            anchor_vector = a_output:clone()
+
+            pos_of_anchor = data.positive[anchor_name]
+            if pos_of_anchor ~= nil then
+                local dupcheck = {}
+                local batch_name = {}
+                minbatchSize = math.min(#pos_of_anchor,4)
+                nsz[1] = minbatchSize
+                batch:resize(nsz)
+                for pi=1,minbatchSize do
+                    positive_name = pos_of_anchor[math.random(#pos_of_anchor)]
+                    if dupcheck[positive_name] == nil then
+                        dupcheck[positive_name] = true
+                        local img = LoadNormalizedResolutionImage(positive_name) 
+                        table.insert(batch_name, positive_name)
+                        batch[pi]:copy(img)
+                    end
+                end
+
+                positive_vector = embedding_net:forward( batch )
+                local max_pdist = -1 
+                local max_im = {}
+                for pi=1,minbatchSize do
+                    bdist = dist(anchor_vector, positive_vector[pi])
+                    --print ("positive_dist", bdist)
+                    if bdist > max_pdist then
+                        max_pdist = bdist
+                        hard_positive_name = batch_name[pi]
+                    end
+                end
+
+                ap_dist = max_pdist
+                if ap_dist ~= 0 then
+                    break
+                end
+            else
+                print ("empty positive", anchor_name)
+            end
+            c1 = math.random(#candidate_anchor_list)
+        end
+
+        local neg_batch = torch.Tensor():type( TensorType )
+        nsz[1] = 4
+        neg_batch:resize( nsz )
+        bisfound = false
+
+        --print ("anchor_name", anchor_name, "ap_dist:", ap_dist)
+        for trial=1,10 do
+            local neg_batch_name = {}
+            for ni=1,4 do
+                local neg_of_anchor = data.negative[anchor_name]
+                if neg_of_anchor == nil or math.random(2) == 1 then
+                    local n1 = c1
+                    local n3 = math.random(nClasses)
+                    while n3 == n1 do
+                        n3 = math.random(nClasses)
+                    end
+
+                    negative_name = data.anchor_name_list[n3]
+                else
+                    local n3 = math.random(#neg_of_anchor)
+                    negative_name = neg_of_anchor[n3]
+                end
+                local img = LoadNormalizedResolutionImage(negative_name)
+                table.insert(neg_batch_name, negative_name)
+                neg_batch[ni]:copy(img)
+            end
+
+            negative_vector = embedding_net:forward(neg_batch)
+            local small_dist = 99999
+            semi_hard_negative_name = nil
+            for ni=1,4 do
+                bdist = dist(anchor_vector, negative_vector[ni])
+                --print ("negative bdist", bdist)
+                if bdist > 0 then
+                    if (bdist > ap_dist and bdist < small_dist) or math.random(5) == 1 then
+                        small_dist = bdist
+                        semi_hard_negative_name = neg_batch_name[ni]
+                    end
+                end
+            end
+
+            if semi_hard_negative_name ~= nil then
+                an_dist = small_dist
+                bisfound = true
+                break
+            end
+        end
+
+        print ("anchor_name", anchor_name, "ap_dist:", ap_dist, "an_dist", an_dist)
+
+        --print(anchor_name, negative_name, positive_name)
+       
+        if bisfound then
+            local exemplar = {anchor_name, semi_hard_negative_name, hard_positive_name}
+            --print ("exemplar", exemplar)
+            table.insert(list, exemplar)
+        end
+    end
+
+    print ("Selection Generate:", #list)
+    return list
+end
+
 function GenerateListTriplets(db, size)
     print ("generate list triplets", size)
-    --print ("data", db)
     local data = db.data
-    --local list = torch.IntTensor(size,3)
     local list = {}
-    -- print ("generate db: ", db)
-    -- print ("Generate List Triplet data:", db)
-    -- print ("anchor_name_list:", data.anchor_name_list)
     local nClasses = #data.anchor_name_list 
     for i=1, size do
         --print ("generate list #" .. i .. "/#" .. size)
         local c1, anchor_name, positive_name, negative_name
+        c1 = i
         while true do
-            c1 = math.random(nClasses)
             anchor_name = data.anchor_name_list[c1]
 
             pos_of_anchor = data.positive[anchor_name]
             if pos_of_anchor ~= nil then
-                --print ("anchor_name", anchor_name)
-                -- print ("pos_list_anchor_name", data.positive)
-                --print ("pos_of_anchor", pos_of_anchor)
                 positive_name = pos_of_anchor[math.random(#pos_of_anchor)]
                 break
             end
+            c1 = math.random(nClasses)
         end
 
         local neg_of_anchor = data.negative[anchor_name]
@@ -75,22 +226,8 @@ function GenerateListTriplets(db, size)
             negative_name = neg_of_anchor[n3]
         end
 
-        --print(anchor_name, negative_name, positive_name)
-       
         local exemplar = {anchor_name, negative_name, positive_name}
         table.insert(list, exemplar)
-
-        --list[i][1] = anchor_name
-        --list[i][2] = negative_name
-        --list[i][3] = positive_name
-
-        if list[i][1] == list[i][3] then
-            print("same index", positive_name, anchor_name)
-        end
-        --print("list1", list[i][1])
-        --print("list2", list[i][2])
-        --print("list3", list[i][3])
-        --print ("generated list:", list)
     end
     return list
 end
@@ -109,16 +246,6 @@ function CreateDistanceTensor(data,labels, model)
         end
     end
     return Dist
-end
-
-
-save_filename = PreProcDir .. '/fashion_data.t7' 
-print ("check chached file:" .. save_filename)
-if path.exists(save_filename) ~= false then
-    print ("cached model")
-    Data = torch.load(save_filename )
-    --print ("Data:", Data.TrainData, Data.TestData)
-    return Data
 end
 
 --function LoadNormalizedResolutionImage(filename)
@@ -169,7 +296,7 @@ function LoadData(filepath)
 
     label_pairs = csvigo.load( {path=DataPath .. filepath, mode='large'} )
 
-    for i=1,#label_pairs do
+    for i=1,#label_pairs,1000 do
         m = label_pairs[i]
         local a_name = m[2]
         local t_name = m[3]
@@ -178,7 +305,7 @@ function LoadData(filepath)
 
         bcontinue = false
 
-        if a_name ~= t_name then
+        if a_name ~= t_name and a_name ~= nil then
             --print ("anchor_name=", a_name)
             if ImagePoolByName[a_name] == nil then
                 local img = LoadNormalizedResolutionImage(a_name)
@@ -226,7 +353,7 @@ function LoadData(filepath)
                 print ("continue")
             end
         end
-        print (i, #label_pairs, 100.0*(i/#label_pairs), count_imagepool)
+        print (i, #label_pairs, 100.0*(i/#label_pairs), count_imagepool, "anchor", #anchor_name_list)
     end
 
     print("loaded: " .. #anchor_name_list)
@@ -238,8 +365,6 @@ function LoadData(filepath)
 
     return Data
 end
-
-local Data = LoadData('fashion_pair.csv')
 
 function SplitData(Data)
     local TrainData = {data={},}
@@ -253,7 +378,8 @@ function SplitData(Data)
     -- split positive
     for i=1, #Data.data.anchor_name_list do
         local anchor_name = Data.data.anchor_name_list[i]
-        if Data.data.positive[anchor_name] ~= nil then
+        local pos_of_anchor = Data.data.positive[anchor_name] 
+        if pos_of_anchor ~= nil then
             if positive_cnt >= 10 then
                 postivie_cnt = 0
                 table.insert(test_anchor_name_list, Data.data.anchor_name_list[i])
@@ -261,9 +387,6 @@ function SplitData(Data)
                 positive_cnt = positive_cnt + 1
                 table.insert(train_anchor_name_list, Data.data.anchor_name_list[i])
             end
-        else
-            positive_cnt = positive_cnt + 1
-            table.insert(train_anchor_name_list, Data.data.anchor_name_list[i])
         end
     end
 
@@ -279,17 +402,64 @@ function SplitData(Data)
     TestData.data.negative = Data.data.negative
     TestData.Resolution = {3, 299, 299}
 
+    print ("Train Data size:", #TrainData.data.anchor_name_list)
+    print ("Test Data size:", #TestData.data.anchor_name_list)
+
     return TrainData, TestData
 end
---TestData = LoadData('fashion_pair_test.csv')
+
+function save_data()
+    torch.save(PreProcDir .. 'fashion_save.t7', 'save')
+    torch.save(PreProcDir .. 'train.resolution.t7', TrainData.Resolution)
+    torch.save(PreProcDir .. 'train.data.anchor_name_list.t7', TrainData.data.anchor_name_list)
+    torch.save(PreProcDir .. 'train.data.positive.t7', TrainData.data.positive)
+    torch.save(PreProcDir .. 'train.data.negative.t7', TrainData.data.negative)
+
+    torch.save(PreProcDir .. 'test.resolution.t7', TestData.Resolution)
+    torch.save(PreProcDir .. 'test.data.anchor_name_list.t7', TestData.data.anchor_name_list)
+    torch.save(PreProcDir .. 'test.data.positive.t7', TestData.data.positive)
+    torch.save(PreProcDir .. 'test.data.negative.t7', TestData.data.negative)
+end
+
+function load_data()
+    if path.exists( PreProcDir .. 'fashion_save.t7') then
+        return nil
+    end
+    TrainData = {data={},Resolution={}}
+    TrainData.Resolution = torch.load(PreProcDir .. 'train.resolution.t7')
+    TrainData.data.anchor_name_list = torch.load(PreProcDir .. 'train.data.anchor_name_list.t7')
+    TrainData.data.positive = torch.load(PreProcDir .. 'train.data.positive.t7')
+    TrainData.data.negative = torch.load(PreProcDir .. 'train.data.negative.t7')
+
+    TestData = {data={},Resolution={}}
+    TestData.Resolution = torch.load(PreProcDir .. 'test.resolution.t7')
+    TestData.data.anchor_name_list = torch.load(PreProcDir .. 'test.data.anchor_name_list.t7')
+    TestData.data.positive = torch.load(PreProcDir .. 'test.data.positive.t7')
+    TestData.data.negative = torch.load(PreProcDir .. 'test.data.negative.t7')
+
+    return RetData { TrainData = TrainData, TestData = TestData }
+end
+
+save_filename = PreProcDir .. '/fashion_data.t7' 
+
+print ("check chached file:" .. save_filename)
+if path.exists(save_filename) ~= false then
+    print ("cached model")
+    Data = torch.load(save_filename )
+
+    print ("Train Data size:", #Data.TrainData.data.anchor_name_list)
+    print ("Test Data size:", #Data.TestData.data.anchor_name_list)
+    return Data
+end
+
+local Data = LoadData('fashion_pair.csv')
 
 TrainData, TestData = SplitData(Data)
-RetData= {TrainData=TrainData,
-    TestData=TestData
-}
+RetData= {TrainData=TrainData, TestData=TestData}
 
 print ('save:' .. save_filename)
 torch.save( save_filename, RetData)
+save_data()
 
 print ("return")
